refactor(utils): move save_to_s3 function from data_cleaning.py to utils.py
